requirements.txt
aiogram>=3.13.1
aiosqlite>=0.20.0
fastapi>=0.123.5
httpx>=0.27.2
faster-whisper>=1.1.1
pydantic>=2.9.2
pydantic-settings>=2.6.1
python-dotenv>=1.0.1
sqlalchemy>=2.0.36
uvicorn>=0.32.0
pytest>=8.3.3
pytest-asyncio>=0.24.0
